requirements.txt
streamlit
pandas
dhanhq
psycopg2-binary
